+++ src/events/Monitor_Handler.js
@@ -1,0 +1,12 @@
+const Event = require('../structures/Event');
+
+module.exports = class extends Event {
+  get options() {
+    return { key: 'message' };
+  }
+
+  async run(message) {
+    this.client._private.monitors.forEach(e => e._run(message));
+    this.client.monitors.forEach(e => e._run(message));
+  }
+};