+++ src/monitors/Command_Handler.js
@@ -1,12 +1,6 @@
-const Event = require('../structures/Event');
+const Monitor = require('../structures/Monitor');
 
-module.exports = class extends Event {
-  get options() {
-    return {
-      key: 'message',
-    };
-  }
-
+module.exports = class extends Monitor {
   async run(message) {
     if (!message) return;
     if (!message.channel) return;
@@ -51,7 +45,7 @@ module.exports = class extends Event {
     if (ignoreCase) cmd = cmd.toLowerCase();
     const filter = e => e.key === cmd || e.aliases.includes(cmd);
     const command = this.client.commands.find(filter);
-    if (!command) return;
+    if (!command) return this.client.monitors.forEach(e => e._run(message));
     const permTest = await this.client.permLevels.test(
       command.permLevel,
       message,