+++ src/structures/Monitor.js
@@ -1,0 +1,15 @@
+const Base = require('./Base');
+
+const defaultOptions = {};
+
+module.exports = class Monitor extends Base {
+  constructor(client, store, fullpath, options = {}) {
+    super(client, store, 'monitor', fullpath, options);
+    /**
+     * @name Monitor#_options
+     * @type {Object}
+     * @private
+     */
+    this._options = { ...defaultOptions, ...this._options };
+  }
+};